requirements.txt
django
djangorestframework
python-dotenv
cloudinary
pytest-monitor